requirements.txt
asgiref==3.8.1
autopep8==2.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.13
djangorestframework==3.15.1
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycodestyle==2.11.1
requests==2.32.3
sqlparse==0.5.0
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
